+++ rlberry_scool/random_agent.py
@@ -1,3 +1,17 @@
-class RandomAgent():
-    def __init__(self):
-        pass
+from rlberry.agents import AgentWithSimplePolicy
+
+# Create random agent as a baseline
+class RandomAgent(AgentWithSimplePolicy):
+    name = "RandomAgent"
+
+    def __init__(self, env, **kwargs):
+        AgentWithSimplePolicy.__init__(self, env, **kwargs)
+
+    def fit(self, budget=100, **kwargs):
+        observation, info = self.env.reset()
+        for ep in range(budget):
+            action = self.policy(observation)
+            observation, reward, done, _ = self.env.step(action)
+
+    def policy(self, observation):
+        return self.env.action_space.sample()  # choose an action at random
